refactor(frontend): dismantle ProductivityScreen God Class
- Extracts complex filtering and state management logic into `useProductivityViewModel` hook to adhere to SRP and decouple logic from UI.
- Extracts `EventCard` into a dedicated component.
- Extracts large `ListEmptyComponent` UI widget tree into `ProductivityEmptyState`.
- Updates `ProductivityScreen` to consume the hook and components.
- Replaces static `T` object colors with the dynamic `useTheme` hook to fully support Dark Mode.
